requirements.txt
jax>=0.1.75
Pillow>=7.2.0